Removed AngularJS module encapsulation
Possibly stupid move, but modules are supposed to make maintenance easier, and the code easier to understand, but I cannot say that was the case here. I haven't even transferred the unit tests from the old repo, which were a big reson for migration to modules in the first place, and AFAIR they weren't truly encapsulated anyways. It's still JavaScript...